Split up big view file
